Massive WFO Update (Part 1 of 3)
Weather Forecasting Office NOAA update. Formatting data as a json file in a resources folder in order to make it easier for the program to load data. It will load the same way as encoding it in one python file. This should make it easier to edit and provide support for the script in the future. 🦭🦭